Changing default value of link text
The previous link color was giving contrast errors with
the WAVE accessibility tool. This change keeps it bright enough to
identify as blue and also comply with WCAG AA and AAA standards.

I preferred this instead of overriding in the accessibility css
(app/assets/stylesheets/hyrax/_accessibility.scss) with a bias
towards fixing the issue at source.
Also this change will improve contrast for all links on the site.